MAINT: Move exceptions from core._internal to core._exceptions
